refactor: remove dead glass debug code, deduplicate card styles, promote HomeBg
- Delete glassStyles.ts, glassStyleStore.ts, and GlassStylePicker.tsx —
  all three were orphaned after the Rim style was hardcoded inline on the
  cards and the picker was removed from PageWrapper
- Extract RIM_OVERLAY, CARD_SHADOW, and CARD_SHADOW_SELECTED from both
  card files into a shared SponsorshipCard/cardStyles.ts so the values
  live in one place
- Move src/components/debug/HomeBgPicker.tsx to src/components/HomeBg.tsx
  now that the picker UI is gone and the component is just the page
  background; update the import in page.tsx accordingly; remove the now-
  empty debug/ directory
