+++ src/app/sponsorships/components/SponsorshipCard/cardStyles.ts
@@ -1,0 +1,11 @@
+/** Shared visual constants for all beneficiary card variants. */
+
+/** Crisp white rim highlight at the top edge of every card photo. */
+export const RIM_OVERLAY =
+  "linear-gradient(to bottom, rgba(255,255,255,0.62) 0px, rgba(255,255,255,0.28) 2px, rgba(255,255,255,0.06) 14px, transparent 32px)"
+
+export const CARD_SHADOW =
+  "0 2px 8px rgba(0,0,0,0.07), 0 6px 24px rgba(0,0,0,0.05)"
+
+export const CARD_SHADOW_SELECTED =
+  "0 0 0 2px rgba(43,127,249,0.18), 0 2px 8px rgba(0,0,0,0.07), 0 6px 24px rgba(0,0,0,0.05)"
