+++ src/components/HomeBg.tsx
